Saving progress for dealing with the 401 error in CreateTestWindow. Tried separating the SSR CSR for createTestWindow but still got 401 in the frontend request.
