+++ backend/src/main/java/org/mentats/mentat/security/WebSecurityConfig.java
@@ -21,6 +21,12 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 import org.mentats.mentat.security.jwt.AuthEntryPointJwt;
 import org.mentats.mentat.security.jwt.AuthTokenFilter;
 
+// Added for CORS
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.CorsConfigurationSource;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+import java.util.Arrays;
+
 /**
  * Global security method configurer
  * Handles JWT requires and what routes
@@ -89,21 +95,30 @@ public class WebSecurityConfig { // extends WebSecurityConfigurerAdapter {
      */
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-         http.cors(Customizer.withDefaults()).csrf().disable()
-                .exceptionHandling().authenticationEntryPoint(unauthorizedHandler).and()
-                .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS).and()
-                .authorizeRequests().requestMatchers("/api/auth/**").permitAll()
-                .requestMatchers("/api/test/**","/api/grades/**","/api/*","/api/createExam/**",
-                        "/course/**",
-                        "/api/test-window/**")
-
-                .permitAll()
-                .anyRequest().authenticated();
+        //  http.cors(Customizer.withDefaults()).csrf().disable()
+        http.cors(cors -> cors.configurationSource(corsConfigurationSource())).csrf().disable()
+            .exceptionHandling().authenticationEntryPoint(unauthorizedHandler).and()
+            .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS).and()
+            .authorizeRequests()
+            .requestMatchers("/api/**", "/course/**")
+            .permitAll().anyRequest().authenticated();
         // Inject the provider into the http handler
         http.authenticationProvider(authenticationProvider());
         // Shift where this filter is added into the chain
         http.addFilterBefore(authenticationJwtTokenFilter(), UsernamePasswordAuthenticationFilter.class);
 
         return http.build();
+    }
+    @Bean
+    public CorsConfigurationSource corsConfigurationSource() {
+        CorsConfiguration configuration = new CorsConfiguration();
+        configuration.setAllowedOrigins(Arrays.asList("http://localhost:3000"));
+        configuration.setAllowedMethods(Arrays.asList("GET", "POST", "PUT", "DELETE", "OPTIONS"));
+        configuration.setAllowedHeaders(Arrays.asList("*"));
+        configuration.setAllowCredentials(true);
+        
+        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+        source.registerCorsConfiguration("/**", configuration);
+        return source;
     }
 }
